Add solution to Challenge 5: DC circuits with uncertain data
Using MATLAB, build the matrix and the right-hand side of the linear systems necessary to determine the current values in the loops of the DC circuit with loop blocks added to the right of the first loop, assuming now that Add a Gaussian random perturbation to each voltage in the right-hand side that is of the original voltage value. Estimate the a priori and a posteriori error bounds for this scenario in which only the right-hand side is subject to uncertainty. Determine empirically the maximum number of loop blocks that can be added while maintaining an accuracy of at least the results. Try to apply analogous perturbations to the values of the resistances and to compute the a priori estimate in the case of uncertainty on the matrix components only.